Enum iced_x86::MemorySizeOptions

pub enum MemorySizeOptions {
    Default,
    Always,
    Minimal,
    Never,
}

Memory size options used by the formatters

Variants

Default

Show memory size if the assembler requires it, else don’t show anything

Always

Always show the memory size, even if the assembler doesn’t need it

Minimal

Show memory size if a human can’t figure out the size of the operand

Never

Never show memory size

Implementations

impl MemorySizeOptions

pub fn values( ) -> impl Iterator<Item = MemorySizeOptions> + DoubleEndedIterator + ExactSizeIterator + FusedIterator

Iterates over all MemorySizeOptions enum values