iced_x86

Struct ConstantOffsets

pub struct ConstantOffsets { /* private fields */ }

Contains the offsets of the displacement and immediate. Call Decoder::get_constant_offsets() or Encoder::get_constant_offsets() to get the offsets of the constants after the instruction has been decoded/encoded.

Implementations

impl ConstantOffsets

pub const fn displacement_offset(&self) -> usize

The offset of the displacement, if any

pub const fn displacement_size(&self) -> usize

Size in bytes of the displacement, or 0 if thereโs no displacement

pub const fn immediate_offset(&self) -> usize

The offset of the first immediate, if any.

This field can be invalid even if the operand has an immediate if itโs an immediate that isnโt part of the instruction stream, eg. SHL AL,1.

pub const fn immediate_size(&self) -> usize

Size in bytes of the first immediate, or 0 if thereโs no immediate

pub const fn immediate_offset2(&self) -> usize

The offset of the second immediate, if any.

pub const fn immediate_size2(&self) -> usize

Size in bytes of the second immediate, or 0 if thereโs no second immediate

pub const fn has_displacement(&self) -> bool

true if displacement_offset() and displacement_size() are valid

pub const fn has_immediate(&self) -> bool

true if immediate_offset() and immediate_size() are valid

pub const fn has_immediate2(&self) -> bool

true if immediate_offset2() and immediate_size2() are valid