iced_x86

Trait FormatterOptionsProvider

pub trait FormatterOptionsProvider {
    // Required method
    fn operand_options(
        &mut self,
        instruction: &Instruction,
        operand: u32,
        instruction_operand: Option<u32>,
        options: &mut FormatterOperandOptions,
        number_options: &mut NumberFormattingOptions<'_>,
    );
}

Can override options used by a Formatter

Required Methods

fn operand_options( &mut self, instruction: &Instruction, operand: u32, instruction_operand: Option<u32>, options: &mut FormatterOperandOptions, number_options: &mut NumberFormattingOptions<'_>, )

Called by the formatter. The method can override any options before the formatter uses them.

Arguments

instruction: Instruction

operand: Operand number, 0-based. This is a formatter operand and isn’t necessarily the same as an instruction operand.

instruction_operand: Instruction operand number, 0-based, or None if it’s an operand created by the formatter.

options: Options. Only those options that will be used by the formatter are initialized.

number_options: Number formatting options