Struct iced_x86::NumberFormattingOptions

pub struct NumberFormattingOptions<'a> {
    pub prefix: &'a str,
    pub suffix: &'a str,
    pub digit_separator: &'a str,
    pub digit_group_size: u8,
    pub number_base: NumberBase,
    pub uppercase_hex: bool,
    pub small_hex_numbers_in_decimal: bool,
    pub add_leading_zero_to_hex_numbers: bool,
    pub leading_zeros: bool,
    pub signed_number: bool,
    pub displacement_leading_zeros: bool,
}

Gets initialized with the default options and can be overridden by a FormatterOptionsProvider

Fields
prefix: &'a str

Number prefix or an empty string

suffix: &'a str

Number suffix or an empty string

digit_separator: &'a str

Digit separator or an empty string to not use a digit separator

digit_group_size: u8

Size of a digit group or 0 to not use a digit separator

number_base: NumberBase

Number base

uppercase_hex: bool

Use uppercase hex digits

small_hex_numbers_in_decimal: bool

Small hex numbers (-9 .. 9) are shown in decimal

add_leading_zero_to_hex_numbers: bool

Add a leading zero to hex numbers if there’s no prefix and the number starts with hex digits A-F

leading_zeros: bool

If true, add leading zeros to numbers, eg. 1h vs 00000001h

signed_number: bool

If true, the number is signed, and if false it’s an unsigned number

displacement_leading_zeros: bool

Add leading zeros to displacements

Implementations

impl<'a> NumberFormattingOptions<'a>

pub fn with_immediate(options: &'a FormatterOptions) -> Self

Creates options used when formatting immediate values

Arguments

options: Formatter options to use

pub fn with_displacement(options: &'a FormatterOptions) -> Self

Creates options used when formatting displacements

Arguments

options: Formatter options to use

pub fn with_branch(options: &'a FormatterOptions) -> Self

Creates options used when formatting branch operands

Arguments

options: Formatter options to use

pub fn new( options: &'a FormatterOptions, leading_zeros: bool, signed_number: bool, displacement_leading_zeros: bool ) -> Self

Constructor

Arguments

options: Formatter options to use

leading_zeros: Add leading zeros to numbers, eg. 1h vs 00000001h

signed_number: Signed numbers if true, and unsigned numbers if false

displacement_leading_zeros: Add leading zeros to displacements